Updated project proposal
# **Updated Midterm Proposal** 
**Group Name**: Logan Triangle: Health, History, and Home in Philadelphia
**Members:** Olivia Arena, Deja McCauley, Lindsey Morris, Cassie Truong

## **Changes**

- We’ve lost initial group members, so we’ve recalibrated the scope of our project. 
- We have also changed the way we are approaching our topic of Logan Neighborhood. Rather than thinking of the conditions of Logan Neighborhood as independent and fixed attributes, we want to think about community investment and divestment. We want to look at ways in which we can think about how to invest in Logan in ways that challenge historical disinvestment. 
- To do this, we have focused on broader buckets of public risk, housing, employment, and access to parks. We hope that future explorations will look at how to prioritize investments to strengthen community networks, protect residents from displacement while improving housing conditions, ensuring there are opportunities for small businesses, and that there is adequate access to open space and transit. 
- Our presentation follows this order: public risk, housing, unemployment, and access to parks. 

## **Research Question**
In terms of public health, affordable housing, amenities, and open space, what community needs emerge as a priority for North Philadelphia’s Logan neighborhood when using indicators (race, unemployment, food accessibility, open space access, and housing) to compare the area to the city as a whole?

## **Why this Matters**
Within the Logan neighborhood of Philadelphia, there is an area called Logan Triangle, which encompasses 48 vacant acres - making it one of the city’s largest sites of contiguous unused land. This area has been a site of environmental degradation, which has caused physical harm to the surrounding physical environment and likely the health of residents. Since the early 20th century, Logan Triangle’s history has been fraught. Although it was once home to over 5,000 Philadelphians, the instability of the land and other environmental hazards (both human-caused) led to the demolition of homes and displacement of residents. The 2023 American Planning Association’s Student Design Competition, uses Logan Triangle as a case study. Our team is tasked with creating a long-term plan for the area that focuses on public health, open space and nature, affordable housing, inclusive design and amenities, and the neighborhood’s history. There is an opportunity and need to redress the physical, environmental, and emotional harms that have occurred in Logan Triangle and the intent of the project is to envision a site that best serves the surrounding and future community of Logan Triangle. In this course, we plan to use spatial data and GIS to better understand existing conditions and community needs.

## **Spatial Scope**
The subject area will encompass the Logan Neighborhood of Philadelphia, as defined by the Logan Comprehensive Neighborhood Plan. The Neighborhood is bound by Olney Avenue on the North, E. Wister Street and Stenton Avenue on the West, N. 6th Street on the East, and W. Wingohocking Street on the South. 

For this project, we defined six census tracts as making up the “Logan Neighborhood.” These tracts include (FIPS codes): "42101028000", "42101028100", "42101028200", "42101028300", "42101028400", "42101028500." The Logan Triangle lies in the center of the neighborhood; this 48-acre block of land targeted for redevelopment was the site of a subsidence crisis where all residential homes were eventually demolished and its residents displaced. Because the Triangle has been cleared out of housing, there are no residents to include in Census analysis. Because of the absence of data, we extended our study area to look at the broader neighborhood. 

<img width="777" alt="Screen Shot 2023-02-12 at 11 47 18 PM" src="https://user-images.githubusercontent.com/122328335/218554096-d86e18df-77e7-43ed-9031-a283768a7b8b.png">

## **Data Sources**
Our team will use [Social Explorer](https://www.socialexplorer.com/) and [Philadelphia Open Data](https://www.opendataphilly.org/dataset) to analyze the neighborhood around the Logan Triangle and the greater Philadelphia area. Social Explorer will provide census data needed to better understand the community in Logan. Data downloaded from Social Explorer will include housing, demographics, and health. The majority of our data will come from Philadelphia Open Data, which provides datasets on planning/zoning, transportation, parks/recreation, and the environment. The specific data sources include: 

- Social Explorer (ACS 5-Year Estimates 2017-2021)
- OpenDataPhilly (Census Tracts, Affordable Housing Production from DHCD, Complaints against Police, SEPTA Locations API, Parks and Recreation)

## **Conclusion & Expectations** 
Through our research, we expect to gain insights on the type of community needs present for North Philadelphia’s Logan neighborhood residents. Specifically, we would like to gain a quantitative understanding of the access that Logan residents have to public health services, affordable housing, amenities, and open space. Furthermore, we would like to compare the access that Logan residents have in comparison to the city as a whole. In the future, there will also be opportunities to analyze how variables, such as race and income, intersect with Philadelphia resident’s access to services and amenities. Ultimately, we hope to develop a comprehensive spatial data analysis and produce powerful visuals that demonstrate the community needs of Logan, Philadelphia to inform the recommendations proposed for Logan Triangle. 

## **Future Exploration**
As we reframe our study, we want to look at the specific conditions of the housing stock (age, condition), calls to 311 about needed investment, small businesses and other community amenities, and potential expansion of transportation.